Move connection services to common namespace
